+++ forgero-core/src/main/java/com/sigmundgranaas/forgero/core/resource/data/ResourceLoader.java
@@ -3,8 +3,11 @@ package com.sigmundgranaas.forgero.core.resource.data;
 import com.sigmundgranaas.forgero.core.resource.data.v2.data.DataResource;
 
 import java.util.List;
+import java.util.Optional;
 
-@FunctionalInterface
+
 public interface ResourceLoader {
     List<DataResource> load();
+
+    Optional<DataResource> loadResource(String path);
 }