+++ changelog.md
@@ -8,6 +8,7 @@ you have backed up your save.**
 ## Changelog 0.10.3
 
 * Fixed compatability with Roughly enough resources
+* Changed data loading to never load disabled packages
 * Fixed mining levels on tools
 * Added drop to Assembly table
 * Added assembly table to mineable tag